requirements.txt

flask
flask-cors
python-dotenv
openai
langchain
langchain-community
langchain-openai
pymilvus
tiktoken
pypdf
tenacity
requests
pydantic
gunicorn
urllib3<2
